``` feat: Enhance user profile management and data persistence
- Implemented local storage for user profiles using `AppCacheSettingImpl`.
- Added functionality to
``````
 save and retrieve `UserProfile` data to and from local storage.
- Modified `UserProfileResponse` to include new fields: `website`, `industryType`, `department`, `designation`, and `employee`.
- Updated `UserProfile` model with new fields to match the backend response.
- Implemented methods
``````
 `toEducationTable`, `toExperienceTable` and `toUrlTable` to convert network DTOs to local entities.
- Updated `SetUpAccountViewModel` to fetch and update user data, including saving the user profile to local storage.
- Modified the `updateUserDataState` function to save user profile to the local cache.
- Updated the `updateUserPreference` function to save user data from the `UserProfileResponse`.
- Updated `updateUserDataState` function to change `UiState` of the `_userData`.
- Removed `id` from `UserProfile` since there is `userId`
-
``````
 Modified `EducationTable`, `ExperienceTable` and `UrlTable` to have `id` as Long
- Added a disposable effect to clear the cache when the app is disposed.
- Added `UserDataSource` to `SetUpAccountViewModel` constructor.
```

+++ composeApp/src/commonMain/kotlin/com/sdjic/gradnet/data/local/entity/EducationTable.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "education_table")
 data class EducationTable(
-    @PrimaryKey val id: Int = 0,
+    @PrimaryKey val id: Long = 0,
     val schoolName: String = "",
     val degree: String? = null,
     val field: String? = null,
